refactor: 패키지 구조를 도메인과 프레임워크로 분리
- baseball 패키지: 야구 게임 도메인 로직
- framework 패키지: 재사용 가능한 게임 프레임워크
- utils 패키지: 유틸리티 클래스
