--- src/main/java/com/kakao/onboarding/precource/dukeyun/baseball/Game.java
@@ -1,9 +1,0 @@
-package com.kakao.onboarding.precource.dukeyun.baseball;
-
-public interface Game {
-
-	void init();
-
-	void play();
-
-}
