--- src/main/java/com/kakao/onboarding/precource/dukeyun/baseball/GameSystem.java
@@ -1,6 +1,0 @@
-package com.kakao.onboarding.precource.dukeyun.baseball;
-
-public interface GameSystem {
-
-	void run();
-}
