--- src/main/java/com/kakao/onboarding/precource/dukeyun/baseball/TurnInput.java
@@ -1,4 +1,0 @@
-package com.kakao.onboarding.precource.dukeyun.baseball;
-
-public interface TurnInput {
-}
