--- src/main/java/com/kakao/onboarding/precource/dukeyun/baseball/TurnResult.java
@@ -1,7 +1,0 @@
-package com.kakao.onboarding.precource.dukeyun.baseball;
-
-public interface TurnResult {
-
-	boolean isGameCleared();
-
-}
