+++ src/main/java/com/kakao/onboarding/precource/dukeyun/baseball/framework/game/Game.java
@@ -1,0 +1,9 @@
+package com.kakao.onboarding.precource.dukeyun.baseball.framework.game;
+
+public interface Game {
+
+	void init();
+
+	void play();
+
+}
