+++ src/main/java/com/kakao/onboarding/precource/dukeyun/baseball/framework/game/turn/TurnBaseGameView.java
@@ -1,4 +1,4 @@
-package com.kakao.onboarding.precource.dukeyun.baseball;
+package com.kakao.onboarding.precource.dukeyun.baseball.framework.game.turn;
 
 public interface TurnBaseGameView {
