+++ src/main/java/com/kakao/onboarding/precource/dukeyun/baseball/framework/gamesystem/GameSystem.java
@@ -1,0 +1,6 @@
+package com.kakao.onboarding.precource.dukeyun.baseball.framework.gamesystem;
+
+public interface GameSystem {
+
+	void run();
+}
