+++ src/main/java/com/kakao/onboarding/precource/dukeyun/baseball/framework/gamesystem/ReplayableGameSystem.java
@@ -1,4 +1,6 @@
-package com.kakao.onboarding.precource.dukeyun.baseball;
+package com.kakao.onboarding.precource.dukeyun.baseball.framework.gamesystem;
+
+import com.kakao.onboarding.precource.dukeyun.baseball.framework.game.Game;
 
 public class ReplayableGameSystem implements GameSystem {
 
@@ -12,8 +14,8 @@ public class ReplayableGameSystem implements GameSystem {
 
 	@Override
 	public void run() {
-		boolean replay = false;
-		while (!replay) {
+		boolean replay = true;
+		while (replay) {
 			game.init();
 			game.play();
 			replay = systemView.requestReplayInput();
