+++ src/main/java/com/kakao/onboarding/precource/dukeyun/baseball/utils/Console.java
@@ -1,4 +1,4 @@
-package com.kakao.onboarding.precource.dukeyun.baseball;
+package com.kakao.onboarding.precource.dukeyun.baseball.utils;
 
 import java.util.Scanner;
